Schematic Verifier: Fix adding the wrong chunks as pending with Auto Refresh
There is still another issue related to Auto Refresh, where chunks
getting loaded on the client will mark them for re-check, but due to
the adjacent chunk requirement they can't be checked yet. And thus
it's not possible to complete all the pending chunks except for very
carefully going over the placement in an axis aligned linear movement.